% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EunomiaData.R
\name{exportDataFiles}
\alias{exportDataFiles}
\title{Export data files from a database(sqlite or duckdb)}
\usage{
exportDataFiles(
  dbPath,
  outputFolder,
  outputFormat = "csv",
  dbms = "sqlite",
  verbose = FALSE
)
}
\arguments{
\item{dbPath}{The path to the source .sqlite or .duckdb file}

\item{outputFolder}{The path to the export destination directory}

\item{outputFormat}{The output format for the files.  Supported formats include csv, parquet.}

\item{dbms}{The file-based database system to use: 'sqlite' (default) or 'duckdb'}

\item{verbose}{Boolean argument controlling verbose debugging output}
}
\value{
No return value, called to export to outputFolder.
}
\description{
Helper function to export data to csv or parquet files from a database file (sqlite or duckdb).
}
