% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportToCsv}
\alias{exportToCsv}
\title{Extract the Eunomia database to CSV files}
\usage{
exportToCsv(
  outputFolder = file.path(getwd(), "csv"),
  connectionDetails = getEunomiaConnectionDetails()
)
}
\arguments{
\item{outputFolder}{A folder where the CSV files will be written.}

\item{connectionDetails}{Connection details for the Eunomia database. Defaults to a fresh Eunomia
database.}
}
\description{
Extract the Eunomia database to CSV files
}
\examples{
\dontrun{
exportToCsv("c:/temp/csv")
}

}
