% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{getConnectionDetails}
\alias{getConnectionDetails}
\title{Get Eunomia Connection Details}
\usage{
getConnectionDetails(
  datasetName,
  cdmVersion = "5.3",
  pathToData = Sys.getenv("EUNOMIA_DATA_FOLDER"),
  dbms = "sqlite",
  autoDownload = TRUE
)
}
\arguments{
\item{datasetName}{The data set name as found on https://github.com/OHDSI/EunomiaDatasets. The
data set name corresponds to the folder with the data set ZIP files}

\item{cdmVersion}{The OMOP CDM version. This version will appear in the suffix of the data file,
for example: <datasetName>_<cdmVersion>.zip. Default: '5.3'}

\item{pathToData}{The path where the Eunomia data is stored on the file system., By default the
value of the environment variable "EUNOMIA_DATA_FOLDER" is used.}

\item{dbms}{The DBMS to create a connection details object to support.  Default is sqlite.}

\item{autoDownload}{Controls if the CDM zip archive is automatically downloaded if the data is not
currently available.}
}
\value{
A ConnectionDetails object, to be used with the \code{DatabaseConnector} package.
}
\description{
Creates a copy of the default (GiBleed) Eunomia database, and provides details for connecting to
that copy. Function provides backwards compatibility to prior releases of Eunomia default (GiBleed)
dataset
}
