% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{listCdmSamples}
\alias{listCdmSamples}
\title{List available CDM samples}
\usage{
listCdmSamples()
}
\value{
A data frame with two columns. The 'cdm_name' column contains the data sample name,
and 'description' column contains brief description of the sample content
}
\description{
List currently available CDM data samples in the Eumonia package
}
\examples{
listCdmSamples()

}
