% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Construct cohorts}
\usage{
createCohorts(
  connectionDetails,
  cdmDatabaseSchema = "main",
  cohortDatabaseSchema = "main",
  cohortTable = "cohort"
)
}
\arguments{
\item{connectionDetails}{The connection details to connect to the (Eunomia) database.}

\item{cdmDatabaseSchema}{Deprecated. The cdm must be created in the main schema.}

\item{cohortDatabaseSchema}{Deprecated. The cohort table will be created in the main schema.}

\item{cohortTable}{Deprecated. Cohort table will be named "cohort".}
}
\value{
A data frame listing all created cohorts.
}
\description{
Creates a set of predefined cohorts in a cohort table. WARNING: this will delete all existing
cohorts in the table!
}
