% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\docType{package}
\name{Eunomia-package}
\alias{Eunomia}
\alias{Eunomia-package}
\title{Eunomia: A Standard Dataset in the OMOP Common Data Model}
\description{
A standard dataset in the `OMOP` (`Observational Medical Outcomes Partnership`) `Common Data Model`, for testing and demonstration purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/Eunomia}
  \item Report bugs at \url{https://github.com/OHDSI/Eunomia/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Frank DeFalco
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
