% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\docType{package}
\name{Eunomia-package}
\alias{Eunomia}
\alias{Eunomia-package}
\title{Eunomia: A Standard Dataset in the OMOP Common Data Model}
\description{
A sample dataset in the OMOP (Observational Medical Outcomes Partnership) Common Data Model (CDM) format. The CDM enables uniform storage of observational health care data, and is widely used for health care analytics. 'Eunomia' contains simulated data as well as a subset of the OMOP Vocabulary, and enables testing of additional packages and is used for educational and demonstration purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/Eunomia}
  \item Report bugs at \url{https://github.com/OHDSI/Eunomia/issues}
}

}
\author{
\strong{Maintainer}: Frank DeFalco \email{fdefalco@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Anthony Sena \email{sena@ohdsi.org}
  \item Natthawut Adulyanukosol \email{na339@unc.edu}
  \item Star Liu \email{sliu197@jhmi.edu}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
