% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\docType{package}
\name{Eunomia-package}
\alias{Eunomia}
\alias{Eunomia-package}
\title{Eunomia: A Standard Common Data Model Dataset}
\description{
A standard CDM dataset for testing and demonstration purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/Eunomia}
  \item Report bugs at \url{https://github.com/OHDSI/Eunomia/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Frank DeFalco
}

}
\keyword{internal}
