% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eunomia.R
\docType{package}
\name{Eunomia-package}
\alias{Eunomia}
\alias{Eunomia-package}
\title{Eunomia: Standard Dataset Manager for Observational Medical Outcomes Partnership Common Data Model Sample Datasets}
\description{
Facilitates access to sample datasets from the 'EunomiaDatasets' repository (\url{https://github.com/ohdsi/EunomiaDatasets}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/Eunomia}
  \item Report bugs at \url{https://github.com/OHDSI/Eunomia/issues}
}

}
\author{
\strong{Maintainer}: Frank DeFalco \email{fdefalco@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Anthony Sena \email{sena@ohdsi.org}
  \item Natthawut Adulyanukosol \email{na339@unc.edu}
  \item Star Liu \email{sliu197@jhmi.edu}
  \item Adam Black \email{black@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
