% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EunomiaData.R
\name{extractLoadData}
\alias{extractLoadData}
\title{Extract the Eunomia data files and load into a SQLite database
Extract files from a .ZIP file and creates a SQLite OMOP CDM database that is then stored in the
same directory as the .ZIP file.}
\usage{
extractLoadData(dataFileName)
}
\arguments{
\item{dataFileName}{The path to the .ZIP file that contains the data}
}
\description{
Extract the Eunomia data files and load into a SQLite database
Extract files from a .ZIP file and creates a SQLite OMOP CDM database that is then stored in the
same directory as the .ZIP file.
}
\examples{
\dontrun{
extractLoadData("c:/strategusData/GiBleed_5.3.zip")
}
}
\seealso{
\code{\link[Eunomia]{downloadEunomiaData}}
}
