% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{getEunomiaConnectionDetails}
\alias{getEunomiaConnectionDetails}
\title{Get Default Eunomia Connection Details}
\usage{
getEunomiaConnectionDetails()
}
\value{
A ConnectionDetails object, to be used with the \code{DatabaseConnector} package.
}
\description{
Creates a copy of the default (GiBleed) Eunomia database, and provides details for connecting to
that copy. Function provides backwards compatibility to prior releases of Eunomia default (GiBleed)
dataset
}
