% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{getEunomiaConnectionDetails}
\alias{getEunomiaConnectionDetails}
\title{Get Eunomia Connection Details}
\usage{
getEunomiaConnectionDetails()
}
\value{
A ConnectionDetails object, to be used with the \code{DatabaseConnector} package.
}
\description{
Get Eunomia Connection Details
}
\examples{
connectionDetails <- getEunomiaConnectionDetails()
connection <- connect(connectionDetails)
querySql(connection, "SELECT COUNT(*) FROM person;")
disconnect(connection)

}
